refactor get-comment-by-target action
